output consistency: allow providing more fine-grained input-args data to ReturnTypeSpec
