+++ misc/python/materialize/output_consistency/operation/return_type_spec.py
@@ -7,10 +7,20 @@
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
 
+from __future__ import annotations
+
 from materialize.output_consistency.data_type.data_type_category import DataTypeCategory
-from materialize.output_consistency.input_data.return_specs.input_arg_type_hints import (
-    InputArgTypeHints,
-)
+
+
+class InputArgTypeHints:
+    def __init__(self):
+        self.type_category_of_requested_args: dict[int, DataTypeCategory] = dict()
+        """Only filled for specified indices_of_required_input_type_hints"""
+        self.return_type_spec_of_requested_args: dict[int, ReturnTypeSpec] = dict()
+        """Only filled for specified indices_of_required_input_type_hints and if requires_return_type_spec_hints is True"""
+
+    def is_empty(self) -> bool:
+        return len(self.type_category_of_requested_args) == 0
 
 
 class ReturnTypeSpec:
@@ -23,6 +33,7 @@ class ReturnTypeSpec:
     ):
         self.type_category = type_category
         self.indices_of_required_input_type_hints = indices_of_required_input_type_hints
+        self.requires_return_type_spec_hints = False
 
         assert (
             type_category != DataTypeCategory.ANY
